Paragraph component created #14
