--- src/ui/home/About.tsx
@@ -1,19 +1,0 @@
-import React, { FunctionComponent } from "react";
-
- type AboutProps = {
-
- }
-
-export const About = ({}: AboutProps) => {
-    return (
-    <div className="flex flex-col h-[760px]">
-        <div className="">
-        <p className="text-3xl xl:text-[40px] leading-relaxed font-brilon text-calicoGray-400 text-center px-10 xl:px-28 py-20">
-        Once upon there was a south african girly girl named Suané. She was always drawn to light and beauty, so she decided to paint with light to show off the colours this world have to offer. She studied graphic design in vancouver, canada. Never defined by a place or a thing, she always aims to try and be a master of all trades, a jack of none.            
-        </p>
-     </div>
-    
-    </div>
-
-    )
-}
